third_party/blink/web_tests/wpt_internal/css/css-conditional/container-queries/anchored/at-container-anchored-serialization.html

<!DOCTYPE html>
<title>CSS Container Queries: anchored() conditionText serialization</title>
<link rel="help" href="https://drafts.csswg.org/css-conditional-5/#container-queries">
<link rel="help" href="https://drafts.csswg.org/cssom/#serialize-a-css-rule">
<link rel="help" href="https://github.com/w3c/csswg-drafts/issues/8171">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/css-conditional/container-queries/support/cq-testcommon.js"></script>
<style id="testSheet">
  @container anchored(        fallback:2) { }
  @container ANChoreD(fallback:    ) { }
  @container anchoRed(fallback) { }
  @container  anchored(  ( fallback: 0) OR ( FALLBACK = 2  ) ) { }
  @container anchored (fallback: 1) { }
</style>
<script>
  setup(() => {
    assert_equals(testSheet.sheet.cssRules.length, 5);
  });

  const tests = [
      ["anchored(fallback: 2)", "Normalize spaces"],
      ["ANChoreD(fallback:    )", "No value - invalid, serializes as <general-enclosed>"],
      ["anchored(fallback)", "Boolean context"],
      ["anchored((fallback: 0) or (fallback = 2))", "Logical with 'or'"],
      ["anchored (fallback: 1)", "Not an anchored() function with space before '('"]
  ].map((e, i) => [testSheet.sheet.cssRules[i], ...e]);

  tests.forEach((t) => {
    test(() => assert_equals(t[0].conditionText, t[1]), t[2]);
  });
</script>
